+++ src/layers/cip/EPath/segments/index.js
@@ -6,5 +6,5 @@ module.exports = {
   Network: require('./network'),
   Symbolic: require('./symbolic'),
   Data: require('./data'),
-  // DataType: require('./datatype')
+  DataType: require('./datatype')
 };